+++ index.d.ts
@@ -4,6 +4,14 @@ declare module "mongoose" {
   import mongoose = require('mongoose');
   import stream = require('stream');
 
+  export enum ConnectionStates {
+    disconnected = 0,
+    connected = 1,
+    connecting = 2,
+    disconnecting = 3,
+    uninitialized = 99,
+  }
+
   /** The Mongoose Date [SchemaType](/docs/schematypes.html). */
   export type Date = Schema.Types.Date;
 
@@ -49,6 +57,9 @@ declare module "mongoose" {
 
   /** The various Mongoose SchemaTypes. */
   export var SchemaTypes: typeof Schema.Types;
+
+  /** Expose connection states for user-land */
+  export var STATES: typeof ConnectionStates;
 
   /** Opens Mongoose's default connection to MongoDB, see [connections docs](https://mongoosejs.com/docs/connections.html) */
   export function connect(uri: string, options: ConnectOptions, callback: (err: CallbackError) => void): void;
@@ -124,7 +135,7 @@ declare module "mongoose" {
 
   type Mongoose = typeof mongoose;
 
-  interface ClientSession extends mongodb.ClientSession {}
+  interface ClientSession extends mongodb.ClientSession { }
 
   interface ConnectOptions extends mongodb.MongoClientOptions {
     /** Set to false to [disable buffering](http://mongoosejs.com/docs/faq.html#callback_never_executes) on all models associated with this connection. */
@@ -293,12 +304,53 @@ declare module "mongoose" {
     watch(pipeline?: Array<any>, options?: mongodb.ChangeStreamOptions): mongodb.ChangeStream;
   }
 
-  class Collection {
+   /*
+   * section collection.js
+   * http://mongoosejs.com/docs/api.html#collection-js
+   */
+  interface CollectionBase extends mongodb.Collection {
+    /*
+      * Abstract methods. Some of these are already defined on the
+      * mongodb.Collection interface so they've been commented out.
+      */
+    ensureIndex(...args: any[]): any;
+    findAndModify(...args: any[]): any;
+    getIndexes(...args: any[]): any;
+
+    /** The collection name */
+    collectionName: string;
+    /** The Connection instance */
+    conn: Connection;
+    /** The collection name */
     name: string;
+  }
+
+  /*
+   * section drivers/node-mongodb-native/collection.js
+   * http://mongoosejs.com/docs/api.html#drivers-node-mongodb-native-collection-js
+   */
+  let Collection: Collection;
+  interface Collection extends CollectionBase {
+    /**
+     * Collection constructor
+     * @param name name of the collection
+     * @param conn A MongooseConnection instance
+     * @param opts optional collection options
+     */
+    new(name: string, conn: Connection, opts?: any): Collection;
+    /** Formatter for debug print args */
+    $format(arg: any): string;
+    /** Debug print helper */
+    $print(name: any, i: any, args: any[]): void;
+    /** Retrieves information about this collections indexes. */
+    getIndexes(): any;
   }
 
   class Document {
     constructor(doc?: any);
+
+    /** This documents _id. */
+    _id?: any;
 
     /** Don't run validation on this path or persist changes to this path. */
     $ignore(path: string): void;
@@ -815,6 +867,9 @@ declare module "mongoose" {
     writeConcern?: any;
   }
 
+  /** Alias for QueryOptions for backwards compatability. */
+  type ModelUpdateOptions = QueryOptions;
+
   interface SaveOptions {
     checkKeys?: boolean;
     validateBeforeSave?: boolean;
@@ -1044,7 +1099,7 @@ declare module "mongoose" {
   }
 
   interface SchemaDefinition {
-    [path: string]: SchemaTypeOptions<any> | Function | string | Schema | Schema[] | Array<SchemaTypeOptions<any>> | Function[];
+    [path: string]: SchemaTypeOptions<any> | Function | string | Schema | Schema[] | Array<SchemaTypeOptions<any>> | Function[] | SchemaDefinition | SchemaDefinition[];
   }
 
   interface SchemaOptions {
@@ -1200,7 +1255,7 @@ declare module "mongoose" {
      * Determines whether a type set to a POJO becomes
      * a Mixed path or a Subdocument (defaults to true).
      */
-    typePojoToMixed?:boolean;
+    typePojoToMixed?: boolean;
   }
 
   interface SchemaTimestampsConfig {
@@ -1216,7 +1271,7 @@ declare module "mongoose" {
     alias?: string;
 
     /** Function or object describing how to validate this schematype. See [validation docs](https://mongoosejs.com/docs/validation.html). */
-    validate?: RegExp | [RegExp, string] | Function;
+    validate?: RegExp | [RegExp, string] | Function | [Function , string];
 
     /** Allows overriding casting logic for this individual path. If a string, the given string overwrites Mongoose's default cast error message. */
     cast?: string;
@@ -1226,7 +1281,7 @@ declare module "mongoose" {
      * path cannot be set to a nullish value. If a function, Mongoose calls the
      * function and only checks for nullish values if the function returns a truthy value.
      */
-    required?: boolean | (() => boolean);
+    required?: boolean | (() => boolean) | [boolean, string];
 
     /**
      * The default value for this path. If a function, Mongoose executes the function
@@ -1329,10 +1384,10 @@ declare module "mongoose" {
     uppercase?: boolean;
 
     /** If set, Mongoose will add a custom validator that ensures the given string's `length` is at least the given number. */
-    minlength?: number;
+    minlength?: number | [number, string];
 
     /** If set, Mongoose will add a custom validator that ensures the given string's `length` is at most the given number. */
-    maxlength?: number;
+    maxlength?: number | [number, string];
   }
 
   interface IndexOptions {
@@ -1578,7 +1633,7 @@ declare module "mongoose" {
 
     // var objectId: mongoose.Types.ObjectId should reference mongodb.ObjectID not
     //   the ObjectIdConstructor, so we add the interface below
-    interface ObjectId extends mongodb.ObjectID {}
+    interface ObjectId extends mongodb.ObjectID { }
 
     class Subdocument extends Document {
       $isSingleNested: true;
@@ -1706,7 +1761,7 @@ declare module "mongoose" {
 
     /** Creates a `findOneAndRemove` query: atomically finds the given document and deletes it. */
     findOneAndRemove(filter?: FilterQuery<DocType>, options?: QueryOptions | null, callback?: (err: any, doc: DocType | null, res: any) => void): Query<DocType | null, DocType>;
-    
+
     /** Creates a `findOneAndUpdate` query: atomically find the first document that matches `filter` and apply `update`. */
     findOneAndUpdate(filter?: FilterQuery<DocType>, update?: UpdateQuery<DocType>, options?: QueryOptions | null, callback?: (err: any, doc: DocType | null, res: any) => void): Query<DocType | null, DocType>;
 
@@ -1763,7 +1818,7 @@ declare module "mongoose" {
     j(val: boolean | null): this;
 
     /** Sets the lean option. */
-    lean(val?: boolean | any): Query<LeanDocument<DocType>, DocType>;
+    lean(val?: boolean | any): Query<LeanDocumentOrArray<ResultType>, DocType>;
 
     /** Specifies the maximum number of documents the query will return. */
     limit(val: number): this;
@@ -1974,22 +2029,22 @@ declare module "mongoose" {
 
   export type FilterQuery<T> = {
     [P in keyof T]?: P extends '_id'
-      ? [Extract<T[P], mongodb.ObjectId>] extends [never]
-        ? mongodb.Condition<T[P]>
-        : mongodb.Condition<T[P] | string | { _id: mongodb.ObjectId }>
-      : [Extract<T[P], mongodb.ObjectId>] extends [never]
-      ? mongodb.Condition<T[P]>
-      : mongodb.Condition<T[P] | string>;
+    ? [Extract<T[P], mongodb.ObjectId>] extends [never]
+    ? mongodb.Condition<T[P]>
+    : mongodb.Condition<T[P] | string | { _id: mongodb.ObjectId }>
+    : [Extract<T[P], mongodb.ObjectId>] extends [never]
+    ? mongodb.Condition<T[P]>
+    : mongodb.Condition<T[P] | string>;
   } &
     mongodb.RootQuerySelector<T>;
-  
+
   export type UpdateQuery<T> = mongodb.UpdateQuery<DocumentDefinition<T>> & mongodb.MatchKeysAndValues<DocumentDefinition<T>>;
 
   export type DocumentDefinition<T> = Omit<T, Exclude<keyof Document, '_id'>>;
 
   type FunctionPropertyNames<T> = {
     // The 1 & T[K] check comes from: https://stackoverflow.com/questions/55541275/typescript-check-for-the-any-type
-    [K in keyof T]: 0 extends (1 & T[K]) ? never : (T[K] extends Function ? K : never) 
+    [K in keyof T]: 0 extends (1 & T[K]) ? never : (T[K] extends Function ? K : never)
   }[keyof T];
 
   type actualPrimitives = string | boolean | number | bigint | symbol | null | undefined;
@@ -2003,13 +2058,18 @@ declare module "mongoose" {
 
   export type _LeanDocument<T> = {
     [K in keyof T]:
-      0 extends (1 & T[K]) ? T[K] : // any
-      T[K] extends unknown[] ? LeanType<T[K][number]>[] : // Array
-      T[K] extends Document ? LeanDocument<T[K]> : // Subdocument
-      T[K];
+    0 extends (1 & T[K]) ? T[K] : // any
+    T[K] extends unknown[] ? LeanType<T[K][number]>[] : // Array
+    T[K] extends Document ? LeanDocument<T[K]> : // Subdocument
+    T[K];
   };
 
   export type LeanDocument<T> = Omit<Omit<_LeanDocument<T>, Exclude<keyof Document, '_id'>>, FunctionPropertyNames<T>>;
+
+  export type LeanDocumentOrArray<T> = 0 extends (1 & T) ? T :
+    T extends unknown[] ? LeanDocument<T[number]>[] :
+    T extends Document ? LeanDocument<T> :
+    T;
 
   class QueryCursor<DocType extends Document> extends stream.Readable {
     /**
@@ -2238,7 +2298,7 @@ declare module "mongoose" {
       type?: string): this;
   }
 
-  class NativeError extends global.Error {}
+  class NativeError extends global.Error { }
   type CallbackError = NativeError | null;
 
   class Error extends global.Error {
@@ -2316,10 +2376,10 @@ declare module "mongoose" {
 
     export class ValidationError extends Error {
       name: 'ValidationError';
-  
-      errors: {[path: string]: ValidatorError | CastError};
+
+      errors: { [path: string]: ValidatorError | CastError };
     }
-  
+
     export class ValidatorError extends Error {
       name: 'ValidatorError';
       properties: {