+++ test/typescript/global.ts
@@ -7,6 +7,7 @@ mongoose.get('useCreateIndex');
 const m: mongoose.Mongoose = new mongoose.Mongoose();
 
 m.set('useUnifiedTopology', true);
+m.STATES.connected;
 
 m.connect('mongodb://localhost:27017/test').then(() => {
   console.log('Connected!');